+++ wuzei/src/logic/custom_protocol.rs
@@ -4,9 +4,7 @@ use std::process::Stdio;
 use tokio_util::codec::*;
 use futures::prelude::*;
 use std::str::FromStr;
-use numpy::{PyArray, PyArray2, prelude::*};
-use pyo3::{Python, Py, /* types::PyAny */ Bound};
-
+use hraw::prelude::*;
 use huazhi::custom_protocol::{response, response_400, ResContent};
 
 #[derive(Debug, PartialEq, strum::EnumString)]
@@ -16,7 +14,7 @@ pub enum TypeNames {
   Resize,
   Refresh,
   GetPixel,
-  GetPath,
+  // GetPath,
   Env,
   Cd,
   ReadRaw,
@@ -86,62 +84,48 @@ async fn logic(key: TypeNames, payload: serde_json::Value, mutex: Arc<Mutex<huaz
     TypeNames::Resize => {
       let width = payload["width"].as_i64().unwrap_or(320) as usize;
       let height = payload["height"].as_i64().unwrap_or(240) as usize;
-      super::mmf_init(width, height);
+      crate::pixel_init(width, height);
       let dst = serde_json::json!({ "state": "successed" });
       response(ResContent::JSON(&dst))
     }
     TypeNames::Refresh => {
-      use huazhi::memorymappedfile::*;
       println!("{:?}", payload);
       let shift = payload["bitshift"].as_i64().unwrap_or(0) as i32;
       let color = payload["color"].as_i64().unwrap_or(0) as i32;
       let matrix = serde_json::from_value::<[[f64; 3]; 3]>(payload["matrix"].to_owned()).ok();
 
-      let mut mmf = super::get_mmf().lock().unwrap();
-      let value = mmf.get_value();
-      let width = value["width"].as_u64().unwrap() as usize;
-      let height = value["height"].as_u64().unwrap() as usize;
-
-      let acc = mmf.to_accessor();
-      response(ResContent::IMAGE(hraw::processing::slice_to_png(acc.to_slice::<i32>(), width, height, shift, matrix, color)))
+      let pixel = crate::get_pixel().lock().unwrap();
+      let width = pixel.width;
+      let height = pixel.height;
+      response(ResContent::IMAGE(hraw::processing::slice_to_png(pixel.data.as_slice(), width, height, shift, matrix, color)))
     },
     TypeNames::GetPixel => {
-      use huazhi::memorymappedfile::*;
-      let mut mmf = super::get_mmf().lock().unwrap();
-
-      let mmf_value = mmf.get_value();
-      let width = mmf_value["width"].as_u64().unwrap() as i32;
-      let height = mmf_value["height"].as_u64().unwrap() as i32;
-
+      let pixel = crate::get_pixel().lock().unwrap();
+      let width = pixel.width as i32;
+      let height = pixel.height as i32;
       let x = payload["x"].as_i64().unwrap_or(0) as i32;
       let y = payload["y"].as_i64().unwrap_or(0) as i32;
 
       let dst = if x < 0 || width - 1 < x || y < 0 || height - 1 < y {
         serde_json::json!({ "status": "failed", "x" : x, "y" : y, "data": null })
       } else {
-        let pos = 4 * (x + y * width) as usize;
-        let data = mmf.to_accessor().read::<i32>(pos);
+        let data = pixel.data[(x + y * width) as usize];
         serde_json::json!({ "status": "successed", "x" : x, "y" : y, "data": data })
       };
       // let dst = serde_json::json!({ "state": "successed", "payload": path });
       response(ResContent::JSON(&dst))
     },
-    TypeNames::GetPath => {
-      let path = super::get_mmf().lock().unwrap().get_path();
-      let dst = serde_json::json!({ "status": "successed", "payload": path });
-      response(ResContent::JSON(&dst))
-    },
+    // TypeNames::GetPath => {
+    //   let pixel = crate::get_pixel().lock().unwrap();
+    //   let dst = serde_json::json!({ "status": "successed", "payload": pixel.path });
+    //   response(ResContent::JSON(&dst))
+    // },
     TypeNames::ReadRaw => {
-      use hraw::{*, buffer::*};
-      use huazhi::memorymappedfile::*;
       let path = payload["path"].as_str().unwrap_or("");
-      
-      let mut hraw = hraw::Hraw::new(path).unwrap();
-      // hraw.info().unwrap();
-      let size = hraw.header().to_size();
+      let mut hraw = Hraw::new(path).unwrap();
 
-      super::mmf_init(size.0, size.1);
-      let mut mmf = super::get_mmf().lock().unwrap();
+      let (width, height, _) = hraw.header().to_size();
+      let mut pixel = crate::get_pixel().lock().unwrap();
 
       // 数値か文字列か判断
       // (|| { 
@@ -157,59 +141,58 @@ async fn logic(key: TypeNames, payload: serde_json::Value, mutex: Arc<Mutex<huaz
       // })();
 
       let _ = match payload["subpath"].as_i64() {
-        Some(n)=> mmf.to_accessor().to_mut_slice::<i32>().from_hraw(path, n as usize),
-        None => mmf.to_accessor().to_mut_slice::<i32>().from_hraw(path, payload["subpath"].as_str().unwrap_or("data.raw"))
+        Some(n)=> {
+          pixel.width = width;
+          pixel.height = height;
+          pixel.data = hraw.to_vec_i32(n as usize).unwrap();
+        },
+        None => {
+          let subpath = payload["subpath"].as_str().unwrap_or("data.raw");
+          pixel.width = width;
+          pixel.height = height;
+          pixel.data = hraw.to_vec_i32(subpath).unwrap();
+        }
       };
 
       let dst = serde_json::json!({ "status": "successed", "payload": path });
       response(ResContent::JSON(&dst))
     },
     TypeNames::ReadPng  => {
-      use hraw::*;
-      use huazhi::memorymappedfile::*;
       let path = payload["path"].as_str().unwrap_or("");
-      let (width, height) = read_png_head(path);
+      let (width, height, data) = read_png_i32(path).unwrap();
 
-      super::mmf_init(width, height);
-      let mut mmf = super::get_mmf().lock().unwrap();
-      let _ = mmf.to_accessor().to_mut_slice::<i32>().from_png(path);
+      let mut pixel = crate::get_pixel().lock().unwrap();
+      pixel.width = width;
+      pixel.width = height;
+      pixel.data = data;
 
-      let dst = serde_json::json!({ "status": "successed", "payload": path });
+      let dst = serde_json::json!({ "status": "successed", "payload": "path" });
       response(ResContent::JSON(&dst))
     },
     TypeNames::Py1 => {
-      use pyo3::prelude::*;
       let args = payload.as_str().unwrap_or("null");
-      let e : Result<String, pyo3::PyErr> = Python::with_gil(|py: Python| {
-        let sys = py.import_bound("sys")?;
-        sys.setattr("stdout", LoggingStdout.into_py(py))?;
-        let dst = py.eval_bound(args, None, None)?;
-        // let _ = mutex.lock().unwrap().send_event(huazhi::event_handler::UserEvent::TerminalMessage(format!("{dst}\r\n")));
-        Ok(format!("{dst}"))
+
+      let stdout = PythonStdout::new(|msg:&str| {
+        println!("{} {:?}", "stdout from python".magenta(), msg);
+        let _ = crate::EVENT_LOOP.get().unwrap().lock().unwrap().send_event(huazhi::event_handler::UserEvent::TerminalMessage(msg.to_string().replace("\n", "\r\n")));    
       });
-      match e {
+      match eval_py1(args, stdout) {
         Ok(n) => response(ResContent::JSON(&serde_json::json!({ "status" : "succeeded", "data" : n }))),
         Err(_) => response(ResContent::JSON(&serde_json::json!({ "status" : "failed" }))),
       }
     },
     TypeNames::Py => {
-      use pyo3::prelude::*;
       let code = payload["code"].as_str().unwrap_or("null");
       let entry = payload["entry"].as_str().unwrap_or("main");
+      let args : serde_json::Value = serde_json::from_str(payload["args"].as_str().unwrap_or("{}")).unwrap();
+      let pixel = crate::get_pixel().clone();
 
-      let e : Result<String, pyo3::PyErr> = Python::with_gil(|py| {
-        let sys = py.import_bound("sys")?;
-        sys.setattr("stdout", LoggingStdout.into_py(py))?;
-
-        let module = PyModule::from_code_bound(py, code,"", "",)?;
-        module.add("Pixel", Pixel)?;
-        let func = module.getattr(entry)?;
-        let args = py.eval_bound(payload["args"].as_str().unwrap_or("{}"), None, None)?;
-        let dst = func.call1((args,))?; //.extract::<String>()?;
-        Ok(format!("{dst}"))
+      let py : anyhow::Result<serde_json::Value> = pixel.modify_with_py_stdout(entry, code, args, |msg|{  
+        println!("{} {:?}", "stdout from python".magenta(), msg);
+        let _ = crate::EVENT_LOOP.get().unwrap().lock().unwrap().send_event(huazhi::event_handler::UserEvent::TerminalMessage(msg.to_string().replace("\n", "\r\n")));    
       });
-      match e {
-        Ok(n) => response(ResContent::JSON(&serde_json::json!({ "status" : "succeeded", "data" : n }))),
+      match py {
+        Ok(n) => response(ResContent::JSON(&serde_json::json!({ "status" : "succeeded", "detail" : n }))),
         Err(e) => response(ResContent::JSON(&serde_json::json!({ "status" : "failed", "detail" : format!("{e}") }))),
       }
     },
@@ -223,7 +206,7 @@ async fn logic(key: TypeNames, payload: serde_json::Value, mutex: Arc<Mutex<huaz
         .expect("failed to start sleep");
 
       let stdout = child.stdout.take().unwrap();
-      let mut reader = FramedRead::new(stdout, MyDecoder);
+      let mut reader = FramedRead::new(stdout, super::MyDecoder);
 
       while let Some(line) = reader.next().await {
         let a = line.unwrap();
@@ -232,111 +215,7 @@ async fn logic(key: TypeNames, payload: serde_json::Value, mutex: Arc<Mutex<huaz
       }
       response(ResContent::JSON(&serde_json::json!({ "status" : "succeeded"})))
     },
-    _ => {
-      response_400(ResContent::TXT("unknown"))
-    }
+    _ => { response_400(ResContent::TXT("unknown")) }
   };
   Ok(dst)
-}
-
-struct MyDecoder;
-
-impl Decoder for MyDecoder {
-  type Item = String;
-  type Error = std::io::Error;
-
-  fn decode(&mut self, src: &mut bytes::BytesMut) -> Result<Option<Self::Item>, Self::Error> {
-    if let Some(i) = src.iter().position(|&b| b == b'\n') {
-      let line = src.split_to(i + 1);
-      Ok(Some(String::from_utf8_lossy(&line).to_string()))
-    } else {
-      Ok(None)
-    }
-  }
-}
-
-#[pyo3::prelude::pyclass]
-struct LoggingStdout;
-
-#[pyo3::prelude::pymethods]
-impl LoggingStdout {
-  fn write(&self, data: &str) {
-    println!("{} {:?}", "stdout from python".magenta(), data);
-    let _ = crate::PX.get().unwrap().lock().unwrap().send_event(huazhi::event_handler::UserEvent::TerminalMessage(data.to_string().replace("\n", "\r\n")));
-  }
-}
-
-#[pyo3::prelude::pyclass]
-struct Pixel;
-
-#[pyo3::prelude::pymethods]
-impl Pixel {
-  fn get(&self, x:usize, y:usize) -> i32 {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let width = mmf_value["width"].as_u64().unwrap() as usize;
-    let dst = mmf.to_accessor().read::<i32>(4 * (x + y * width)).unwrap();
-    dst
-  }
-  fn set(&self, x:usize, y:usize, src:i32) {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let width = mmf_value["width"].as_u64().unwrap() as usize;
-    mmf.to_accessor().write::<i32>(4 * (x + y * width), src);
-  }
-
-  fn width(&self) -> i32 {
-    let mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let width = mmf_value["width"].as_u64().unwrap() as i32;
-    width
-  }
-  fn height(&self) -> i32 {
-    let mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let height = mmf_value["height"].as_u64().unwrap() as i32;
-    height
-  }
-  fn to_array(&self) -> Vec<i32> {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let width = mmf_value["width"].as_u64().unwrap() as usize;
-    let height = mmf_value["height"].as_u64().unwrap() as usize;
-    let mut dst = vec![0i32; width * height];
-    mmf.to_accessor().read_array::<i32>(0, &mut dst);
-    dst
-  }  
-  fn to_np<'py>(&self, py: Python<'py>) -> Py<PyArray2<i32>> {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    let mmf_value = mmf.get_value();
-    let width = mmf_value["width"].as_u64().unwrap() as usize;
-    let height = mmf_value["height"].as_u64().unwrap() as usize;
-    let mut v = vec![0i32; width * height];
-    mmf.to_accessor().read_array::<i32>(0, &mut v);
-    let arr = PyArray::from_vec_bound(py, v);
-    let pyarray = arr.reshape([height, width]).unwrap();
-    pyarray.into()
-  }
-  fn from_array(&self, src:Vec<i32>) {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    // let mmf_value = mmf.get_value();
-    // let width = mmf_value["width"].as_u64().unwrap() as usize;
-    // let height = mmf_value["height"].as_u64().unwrap() as usize;
-    mmf.to_accessor().write_array::<i32>(0, &src);
-  }  
-  fn from_np<'py>(&self, src: &Bound<'py, PyArray2<i32>>) {
-    use huazhi::memorymappedfile::*;
-    let mut mmf = super::get_mmf().lock().unwrap();
-    // let mmf_value = mmf.get_value();
-    // let width = mmf_value["width"].as_u64().unwrap() as usize;
-    // let height = mmf_value["height"].as_u64().unwrap() as usize;
-    let vec = src.to_vec().unwrap();
-    mmf.to_accessor().write_array::<i32>(0, &vec);
-  }
-
 }